+++ FastAnimations/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Fixed the 'fade' animation speed affecting fades in events.
+* Fixed features breaking if an integration with another mod fails.
 
 ## 1.15.2
 Released 03 September 2025 for SMAPI 4.2.0 or later.